#7 - using a DOM element
Now using a created DOM element to check what should happen to the menu. The DOM eleemnt is called `.rm-mobile-indicator` and the code checks wether the z-index of this element is set to 0 or 1. If set to 0, the mobile menu should be shown, otherwise the desktop menu. The z-inex cahnges together with all the other CSS elements by using the same media querie width.